+++ src/Sylves/Grid/Square/SquareGrid.cs
@@ -344,8 +344,8 @@ namespace Sylves
                 {
                     minCell.x = Math.Max(minCell.x, bound.min.x);
                     minCell.y = Math.Max(minCell.y, bound.min.y);
-                    maxCell.x = Math.Min(minCell.x, bound.max.x - 1);
-                    maxCell.y = Math.Min(minCell.y, bound.max.y - 1);
+                    maxCell.x = Math.Min(maxCell.x, bound.max.x - 1);
+                    maxCell.y = Math.Min(maxCell.y, bound.max.y - 1);
                 }
 
                 // Loop over cels