+++ docs/articles/release_notes.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Fixed bug with square and cube grid bounds.
+
 # 0.1
 
 * Initial public release